Set correct permissions when creating directory
Currently we only grant the current user write, read and execute access
on our target and pages directory. As indicated by the NSFN documentation
the necessary permissions for directories resolve to 0755 for
directories (user has read, write and execute permission; the user group
and others only read and execute). In order to access our files the
execute permission is critical. Without it trying to access pages
residing inside a directory missing it results in a 403.